refactor classes to Board file
