+++ src/data_access/UserDAO.java
@@ -7,6 +7,9 @@ import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.Updates;
 import com.mongodb.client.result.UpdateResult;
 import entity.*;
+import entity.board.EasyBoard;
+import entity.board.GameState;
+import entity.board.HardBoard;
 import org.bson.Document;
 import org.bson.conversions.Bson;
 import org.bson.types.ObjectId;
