--- src/entity/HardBoardFactory.java
@@ -1,6 +1,0 @@
-package entity;
-
-public class HardBoardFactory implements BoardFactory{
-    @Override
-    public Board create() {return new HardBoard();}
-}
