+++ src/entity/board/Board.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.board;
 
 import java.util.HashMap;
