+++ src/entity/board/BoardFactory.java
@@ -1,4 +1,6 @@
-package entity;
+package entity.board;
+
+import entity.board.Board;
 
 public interface BoardFactory {
     public Board create();
