+++ src/entity/board/EasyBoardFactory.java
@@ -1,4 +1,8 @@
-package entity;
+package entity.board;
+
+import entity.board.Board;
+import entity.board.BoardFactory;
+import entity.board.EasyBoard;
 
 public class EasyBoardFactory implements BoardFactory {
     @Override
