+++ src/entity/board/EasySudokuScrambler.java
@@ -1,21 +1,21 @@
-package entity;
+package entity.board;
 
 import java.util.HashMap;
 
-public class HardSudokuScrambler implements SudokuScrambler{
-    private HardBoard currBoard;
+public class EasySudokuScrambler implements SudokuScrambler {
+    private EasyBoard currBoard;
 
-    public HardSudokuScrambler(HardBoard currBoard) {
+    public EasySudokuScrambler(EasyBoard currBoard) {
         this.currBoard = currBoard;
     }
 
-    public HardBoard scramble() {
+    public EasyBoard scramble() {
         HashMap<Integer, Boolean>[][] currBoardValues = currBoard.getCurrBoard();
-        HardBoard newBoard = new HardBoard();
+        EasyBoard newBoard = new EasyBoard();
         int[][] newBoardSolutions = newBoard.getSolutionBoard();
         HashMap<Integer, Boolean>[][] newBoardValues = newBoard.generateBlankBoard();
-        for (int i = 0; i < 9; i++) {
-            for (int j = 0; j < 9; j++) {
+        for (int i = 0; i < 4; i++) {
+            for (int j = 0; j < 4; j++) {
                 if (!currBoardValues[i][j].isEmpty()) {
                     newBoardValues[i][j].put(newBoardSolutions[i][j], true);
                 }
@@ -24,4 +24,5 @@ public class HardSudokuScrambler implements SudokuScrambler{
         newBoard.setBoard(newBoardValues);
         return newBoard;
     }
+
 }
