+++ src/entity/board/HardSudokuScrambler.java
@@ -1,22 +1,21 @@
-package entity;
+package entity.board;
 
 import java.util.HashMap;
-import java.util.Map;
 
-public class EasySudokuScrambler implements SudokuScrambler {
-    private EasyBoard currBoard;
+public class HardSudokuScrambler implements SudokuScrambler {
+    private HardBoard currBoard;
 
-    public EasySudokuScrambler(EasyBoard currBoard) {
+    public HardSudokuScrambler(HardBoard currBoard) {
         this.currBoard = currBoard;
     }
 
-    public EasyBoard scramble() {
+    public HardBoard scramble() {
         HashMap<Integer, Boolean>[][] currBoardValues = currBoard.getCurrBoard();
-        EasyBoard newBoard = new EasyBoard();
+        HardBoard newBoard = new HardBoard();
         int[][] newBoardSolutions = newBoard.getSolutionBoard();
         HashMap<Integer, Boolean>[][] newBoardValues = newBoard.generateBlankBoard();
-        for (int i = 0; i < 4; i++) {
-            for (int j = 0; j < 4; j++) {
+        for (int i = 0; i < 9; i++) {
+            for (int j = 0; j < 9; j++) {
                 if (!currBoardValues[i][j].isEmpty()) {
                     newBoardValues[i][j].put(newBoardSolutions[i][j], true);
                 }
@@ -25,5 +24,4 @@ public class EasySudokuScrambler implements SudokuScrambler {
         newBoard.setBoard(newBoardValues);
         return newBoard;
     }
-
 }
